requirements.txt
python-telegram-bot==20.7
pandas>=2.0.0
python-dateutil>=2.8.2
